Fix final maintainability issues - remove unnecessary casts and use constants
- Removed unnecessary long casts in Thread.sleep() calls:
  * MetricsController: Removed (long) cast from (random.nextInt(100) + 50) and (random.nextInt(1000) + 500)
  * PrometheusController: Removed (long) cast from (random.nextInt(10) + 5)
  * Thread.sleep() accepts long parameter, Java automatically widens int to long

- Used MESSAGE_KEY constant instead of duplicated literals in AuthTestController:
  * cacheAdminEndpoint: Used MESSAGE_KEY for 'Cache admin operation successful'
  * multiPermissionEndpoint: Used MESSAGE_KEY for 'Multi-permission endpoint accessed successfully'
  * allPermissionsEndpoint: Used MESSAGE_KEY for 'All permissions endpoint accessed successfully'

- All tests continue to pass (22/22)
- Eliminated code redundancy and improved maintainability
- Cleaner, more consistent code following best practices

+++ src/main/java/hello/MetricsController.java
@@ -35,7 +35,7 @@ public class MetricsController {
         Timer.Sample sample = Timer.start();
         try {
             // Simulate some work
-            Thread.sleep((long) (random.nextInt(100) + 50));
+            Thread.sleep(random.nextInt(100) + 50);
             return "Metrics test completed";
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
@@ -58,7 +58,7 @@ public class MetricsController {
         Timer.Sample sample = Timer.start();
         try {
             // Simulate slow operation
-            Thread.sleep((long) (random.nextInt(1000) + 500));
+            Thread.sleep(random.nextInt(1000) + 500);
             return "Slow operation completed";
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
